ai/requirements.txt
flask==2.0.3
google-cloud-storage==2.5.0
transformers==4.26.0
torch==1.13.1